Fully working with lots of repeated code

+++ Models/NumberModel.cs
@@ -4,7 +4,7 @@ namespace codeTest.Models
 {
     public class NumberModel
     {
-        public int StartingNumber { get; set; }
+        public decimal StartingNumber { get; set; }
         public string NumberString { get;set; }
         
     }
